requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain Framework
langchain>=0.0.340
langchain-openai>=0.0.2
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5

# Document Processing
PyPDF2>=3.0.1
python-docx>=0.8.11
docx2txt>=0.8

# Vector Databases
pinecone-client>=2.2.4
chromadb>=0.4.15

# OpenAI (for embeddings and models)
openai>=1.3.0

# Anthropic Claude
anthropic>=0.7.0

# Google AI
google-generativeai>=0.3.0

# Firebase (optional)
firebase-admin>=6.2.0

# Text Processing
tiktoken>=0.5.1
sentence-transformers>=2.2.2

# Utilities
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0

# Optional: For advanced features
faiss-cpu>=1.7.4  # Alternative to vector stores
nltk>=3.8.1       # Text processing
spacy>=3.7.0      # NLP processing

# Development Dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0